social_media_backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo;python_version<"3.8"
Django==4.2.1
gunicorn==20.1.0
sqlparse==0.4.4
typing_extensions==4.6.2
whitenoise==6.4.0
python-decouple==3.8
psycopg2-binary==2.9.6
dj-database-url==2.0.0
djangorestframework==3.14.0